Clicks 'Share event' button
Actually sharing the public YouTube link using the form is going to be a challenge. It seems that when I try to add `message` to the 'Add a comment...' field, the interpolated `link` is invoking the `youtube_link` method again. But those DOM elements no longer exist on the page, causing an 'Element is no longer attached to the DOM (Selenium::WebDriver::Error::StaleElementReferenceError)' error. I think I need to figure out a way to save the YouTube link without reinvoking the `youtube_link` method, but I'm having a hard time getting this to work.